extract helper methods
